organisação diretorios

+++ api/mentores/listaMentores.js
@@ -14,12 +14,8 @@ const renderNovo = (listaMentores) =>{
 }
 const editmentor=(mentorid)=>{
     console.log(mentorid)
-    window.location="../mentores/editMentor/editMentor.html?id="+mentorid
+    window.location="../mentores/cadastroEdicaoMentor/edicaoMentor.html?id=" + mentorid
 }
-//const deletar=(mentorid)=>{
-    //console.log(mentorid)
-    
-//}
 
 const getNovo = async () => {
     const response = await fetch("http://localhost:3000/mentores" );
@@ -31,18 +27,18 @@ renderNovo(listaMentoresJson);
 
 }
 const deletar = async (id) => {
-    await fetch('http://localhost:3000/mentores/' + id, {
-        method:'DELETE'              
-    });
-    window.location="../mentores/mentoress.html";
 
-      alert ("o mentor sera excluido")
+    if(confirm("Deseja deletar o item?")){
+        await fetch('http://localhost:3000/mentores/' + id, {
+        method:'DELETE'              
+        });
+        window.location="../mentores/listaMentores.html";
+
+        alert ("O mentor foi excluido!");
+        return;
+    }          
     return;
 }
-
-  
-  
-
 
 getNovo();
