+++ api/mentorias/listaMentorias.js
@@ -8,13 +8,28 @@ const renderMentorias = (mentoria)=>{
             '<tr> ' +
             '<td> ' + mentoria.mentoria + '</td>' +
             '<td> ' +mentoria.mentor + '</td>' + 
-            '<td> ' + mentoria.status + '</td>' +  
-            '<td class="td-action"> <button type ="button" onclick="editmentor('+ mentoria.id+')" >editar</button ><button type ="button" onclick="deletar('+ mentoria.id+')" >Deletar</button ></td>'  
+            //'<td> '  +mentoria.status + '</td>' + 
+            //'<td> ' + (mentoria.status!=null && mentoria.status==true ? "Ativo" : "Inativo") + '</td>' + 
+            '<td> ' + (converterStatus( mentoria.status)) +  '</td>' + 
+            '<td class="td-action"> <button type ="button" onclick="editmentoria('+ mentoria.id+')" >editar</button ><button type ="button" onclick="deletar('+ mentoria.id+')" >Deletar</button ></td>'  
         '</tr>';
         //console.log("mentoria",mentoria)
     
-        });
-   
+        });   
+}
+
+const converterStatus = (statusBoleano) => {
+      //['ativo', 'false'].includes(status) ? status === true : inativo
+      //mentoria.status!=null && mentoria.status==true ? "Ativo" : "Inativo
+      var statusString = "";
+      if (statusBoleano != null && statusBoleano==true){
+        statusString = "Ativo";
+      }else{
+        statusString = "Inativo";
+      }
+
+      return statusString;
+
 }
 /*
  const obterMentor = async (mentorId)=>{
@@ -25,7 +40,7 @@ const renderMentorias = (mentoria)=>{
 */
 const editmentoria=(mentoriaid)=>{
     console.log(mentoriaid)
-    window.location="../mentores/editMentor/editMentor.html?id="+mentoriaid
+    window.location="../mentorias/cadastroEdicaoMentoria/edicaoMentoria.html?id="+mentoriaid
 }
 //const deletar=(mentorid)=>{
     //console.log(mentorid)
@@ -41,13 +56,16 @@ const getMentoria = async () => {
 
 }
 const deletar = async (id) => {
+  if(confirm("Deseja deletar o item?")){
     await fetch('http://localhost:3000/mentorias/' + id, {
         method:'DELETE'              
     });
-    window.location="btn-nova-mentoria";
+    window.location="../mentorias/listaMentorias.html";
 
-      alert ("o mentor sera excluido")
+    alert ("O mentor foi excluido!");
     return;
+  }
+  return;
 }
 getMentoria()
